requirements.txt
streamlit
pandas
langgraph
langchain
langchain-groq
python-dotenv
matplotlib
seaborn
tabulate
